update of the driver initiation

+++ detectionServer/src/main/java/com/detect/util/TagFrame.java
@@ -6,10 +6,10 @@ import org.bytedeco.javacpp.opencv_core;
  * Created by wxy on 2017/12/10.
  */
 public class TagFrame {
-    public double idx;
+    public long idx;
     public opencv_core.Mat origframe;
     public opencv_core.Mat curframe;
-    public TagFrame(opencv_core.Mat origframe,double idx){
+    public TagFrame(opencv_core.Mat origframe,long idx){
         this.origframe = origframe;
         this.idx = idx;
     }
